initial merge with 2020_10

+++ lib/src/phy/rf/rf_dummy_imp.h
@@ -65,9 +65,9 @@ SRSLTE_API   bool   rf_dummy_is_master_clock_dynamic(void *h);
 
 SRSLTE_API   double rf_dummy_set_rx_srate(void *h, double freq);
 
-SRSLTE_API   double rf_dummy_set_rx_gain(void *h, double gain);
+SRSLTE_API   int rf_dummy_set_rx_gain(void *h, double gain);
 
-SRSLTE_API   double rf_dummy_set_tx_gain(void *h, double gain);
+SRSLTE_API   int rf_dummy_set_tx_gain(void *h, double gain);
 
 SRSLTE_API   srslte_rf_info_t * rf_dummy_get_rf_info(void *h);
